+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added base classes for value and  multi value converters and their combination for easy skipping of convert back.
 - Added BooleanToBooleanConverter to convert a single boolean or a list of booleans to another single boolean representation.
+- Added a multi boolean to integer conversion into the BooleanToIntegerConverter.
 ### Removed
 - Removed InverseBooleanConverter, thats replaced by BooleanToBooleanConverter.
 - Removed MultiBooleanToBooleanConverter, thats integrated into the BooleanToBooleanConverter.